Fix cert validation for authorized cluster endpoint on rke clusters
rancher/rancher#16486
rancher/rancher#17793
