+++ lib/shared/addon/components/cluster-driver/driver-rke/component.js
@@ -9,7 +9,7 @@ import {
   underlineToCamel, removeEmpty, keysToCamel, validateEndpoint, randomStr
 } from 'shared/utils/util';
 import { validateHostname } from '@rancher/ember-api-store/utils/validate';
-
+import { validateCertWeakly } from 'shared/utils/util';
 import C from 'shared/utils/constants';
 import YAML from 'yamljs';
 import json2yaml from 'json2yaml';
@@ -436,8 +436,7 @@ export default InputTextFile.extend(ClusterDriver, {
     return out;
   }),
 
-  value: computed('pasteOrUpload', {
-
+  yamlValue: computed('pasteOrUpload', {
     get() {
       const intl = get(this, 'intl');
 
@@ -536,7 +535,7 @@ export default InputTextFile.extend(ClusterDriver, {
   validate() {
     this._super(...arguments);
 
-    let errors = get(this, 'errors') || [];
+    let errors = [];
     let config = get(this, `cluster.rancherKubernetesEngineConfig`);
 
     if ( !get(this, 'isCustom') ) {
@@ -561,6 +560,10 @@ export default InputTextFile.extend(ClusterDriver, {
       errors = this.validateEtcdService(errors);
     }
 
+    if ( get(this, 'cluster.localClusterAuthEndpoint.enabled') ) {
+      errors = this.validateAuthorizedClusterEndpoint(errors);
+    }
+
     const clusterOptErrors = get(this, 'clusterOptErrors') || [];
 
     if ( get(config, 'cloudProvider.name') === 'azure' ) {
@@ -576,6 +579,25 @@ export default InputTextFile.extend(ClusterDriver, {
     set(this, 'errors', errors);
 
     return errors.length === 0 && clusterOptErrors.length === 0;
+  },
+
+  validateAuthorizedClusterEndpoint(errors) {
+    let { localClusterAuthEndpoint } = get(this, 'cluster');
+    let { caCerts, fqdn } = localClusterAuthEndpoint;
+
+    if (caCerts) {
+      if (!validateCertWeakly(caCerts) ) {
+        errors.push(this.intl.t('newCertificate.errors.cert.invalidFormat'));
+      }
+    } else {
+      errors.push(this.intl.t('newCertificate.errors.cert.required'));
+    }
+
+    if (fqdn) {
+      errors = validateHostname(fqdn, 'FQDN', get(this, 'intl'), { restricted: true }, errors);
+    }
+
+    return errors;
   },
 
   validateEtcdService(errors) {
