+++ lib/shared/addon/components/cluster-driver/driver-rke/template.hbs
@@ -49,13 +49,13 @@
               showDownload=false
               canChangeName=false
               autoResize=true
-              value=value
+              value=yamlValue
             }}
           </div>
           {{copy-to-clipboard
             tooltipText=""
             buttonText="copyToClipboard.tooltip"
-            clipboardText=value
+            clipboardText=yamlValue
             class="with-clip"
           }}
         </div>
@@ -706,6 +706,7 @@
 {{top-errors errors=clusterErrors}}
 {{top-errors errors=otherErrors}}
 {{top-errors errors=clusterOptErrors}}
+
 {{#if (or isEdit (eq step 1))}}
   {{save-cancel
     createLabel=(if isCustom "saveCancel.next" "saveCancel.create")
