+++ lib/shared/addon/utils/util.js
@@ -162,11 +162,6 @@ export function absoluteUrl(url) {
   return parseUrl(url).href;
 }
 
-export function validateEndpoint(str) {
-  // credit to https://stackoverflow.com/questions/4460586/javascript-regular-expression-to-check-for-ip-addresses
-  return /^(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$/.test(str);
-}
-
 export function addAuthorization(url, user, pass) {
   url = absoluteUrl(url);
   var pos = url.indexOf('//');
@@ -479,6 +474,35 @@ export function keysToCamel(obj) {
 
 export function deepCopy(obj) {
   return JSON.parse(JSON.stringify(obj));
+}
+
+export function validateEndpoint(str) {
+  // credit to https://stackoverflow.com/questions/4460586/javascript-regular-expression-to-check-for-ip-addresses
+  return /^(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$/.test(str);
+}
+
+export function validateCertWeakly(certs) {
+  const BEGIN_CERTIFICATE = [
+    '-----BEGIN CERTIFICATE-----'
+  ];
+
+  return certs.trim().startsWith(BEGIN_CERTIFICATE[0]);
+}
+
+export function validateKeyWeakly(key) {
+  const BEGIN_KEY = [
+    '-----BEGIN PRIVATE KEY-----',
+    '-----BEGIN EC PRIVATE KEY-----',
+    '-----BEGIN RSA PRIVATE KEY-----',
+  ]
+
+  BEGIN_KEY.forEach((prefix) => {
+    if ( key.trim().startsWith(prefix) ) {
+      return true;
+    }
+  });
+
+  return false;
 }
 
 var Util = {
